requirements.txt
alembic==1.13.1
astroid==3.2.2
bcrypt==4.1.3
blinker==1.8.0
click==8.1.7
colorama==0.4.6
coverage==7.5.3
dill==0.3.8
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg2==2.9.9
pylint==3.2.2
pytest==8.2.2
python-dotenv==1.0.1
SQLAlchemy==2.0.30
tomlkit==0.12.5
typing_extensions==4.11.0
Werkzeug==3.0.2
